+++ src/main/kotlin/com/jetbrains/snakecharm/inspections/SmkUnusedLogFileInspection.kt
@@ -9,6 +9,9 @@ import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
 import com.intellij.psi.util.parentOfType
 import com.intellij.refactoring.suggested.endOffset
+import com.intellij.refactoring.suggested.startOffset
+import com.jetbrains.python.psi.PyBinaryExpression
+import com.jetbrains.python.psi.PyStringElement
 import com.jetbrains.python.psi.PyStringLiteralExpression
 import com.jetbrains.snakecharm.SnakemakeBundle
 import com.jetbrains.snakecharm.codeInsight.SnakemakeAPI.EXECUTION_SECTIONS_THAT_ACCEPTS_SNAKEMAKE_PARAMS_OBJ_FROM_RULE
@@ -17,7 +20,8 @@ import com.jetbrains.snakecharm.lang.psi.*
 
 class SmkUnusedLogFileInspection : SnakemakeInspection() {
     companion object {
-        private const val REDIRECT_STDERR_STDOUT_TO_LOG_CMD_TEXT = " >{log} 2>&1"
+        private const val REDIRECT_STDERR_STDOUT_TO_LOG_CMD_TEXT = ">{log} 2>&1"
+        private const val F_STRING_REDIRECT_STDERR_STDOUT_TO_LOG_CMD_TEXT = ">{{log}} 2>&1"
     }
 
     override fun buildVisitor(
@@ -62,8 +66,11 @@ class SmkUnusedLogFileInspection : SnakemakeInspection() {
             }
 
             val shellSection = rule.getSectionByName(SnakemakeNames.SECTION_SHELL)
+            var hasStringConcatenationInShellSection = false
             val name = rule.name ?: return
             val quickfix = if (shellSection != null) {
+                hasStringConcatenationInShellSection =
+                    shellSection.argumentList?.arguments?.firstOrNull { it is PyBinaryExpression } != null
                 CreateLogFileInShellSection(shellSection, name)
             } else {
                 val runSection = ruleSections.firstOrNull { it is SmkRunSection } as SmkRunSection?
@@ -83,11 +90,10 @@ class SmkUnusedLogFileInspection : SnakemakeInspection() {
                 }
 
             if (!collector.hasReferenceToTarget) {
-                registerProblem(
-                    originalLogSection,
-                    message,
-                    quickfix
-                )
+                when (hasStringConcatenationInShellSection) {
+                    true -> registerProblem(originalLogSection, message)
+                    else -> registerProblem(originalLogSection, message, quickfix)
+                }
             }
         }
 
@@ -155,19 +161,34 @@ class SmkUnusedLogFileInspection : SnakemakeInspection() {
             val shellSection = (startElement as SmkArgsSection)
             val stringArgument = shellSection.argumentList?.arguments?.firstOrNull {
                 it is PyStringLiteralExpression
-            }
-
+            } as? PyStringLiteralExpression
+            val indent = shellSection.getPreviousOffset()?.replace("\n", "")
             if (stringArgument == null) {
-                val indent = shellSection.getPreviousOffset()?.replace("\n", "")
                 doc.insertString(
                     shellSection.lastChild.endOffset,
-                    "\n$indent$indent\"$REDIRECT_STDERR_STDOUT_TO_LOG_CMD_TEXT\""
+                    "\n$indent$indent\" $REDIRECT_STDERR_STDOUT_TO_LOG_CMD_TEXT\""
                 )
                 PsiDocumentManager.getInstance(project).commitDocument(doc)
                 return
             }
-
-            doc.insertString(stringArgument.endOffset - 1, REDIRECT_STDERR_STDOUT_TO_LOG_CMD_TEXT)
+            var fixText = REDIRECT_STDERR_STDOUT_TO_LOG_CMD_TEXT
+            var startOffset = 1
+            val stringElements = stringArgument.stringElements as List<PyStringElement>
+            if (stringElements.firstOrNull()?.isTripleQuoted == true) {
+                startOffset = 3
+            }
+            val endOffset = startOffset
+            if (stringElements.firstOrNull()?.isFormatted == true) {
+                startOffset += 1
+            }
+            if (stringElements.lastOrNull()?.isFormatted == true) {
+                fixText = F_STRING_REDIRECT_STDERR_STDOUT_TO_LOG_CMD_TEXT
+            }
+            doc.insertString(
+                stringArgument.endOffset - endOffset,
+                ") $fixText"
+            )
+            doc.insertString(stringArgument.startOffset + startOffset, "(")
             PsiDocumentManager.getInstance(project).commitDocument(doc)
         }
     }
@@ -186,7 +207,10 @@ class SmkUnusedLogFileInspection : SnakemakeInspection() {
             val runSection = (startElement as SmkRunSection)
             val lastArg = runSection.lastChild
             val indent = lastArg.prevSibling.text
-            doc.insertString(lastArg.endOffset, "${indent}shell(\"echo TODO$REDIRECT_STDERR_STDOUT_TO_LOG_CMD_TEXT\")")
+            doc.insertString(
+                lastArg.endOffset,
+                "${indent}shell(\"(echo TODO) $REDIRECT_STDERR_STDOUT_TO_LOG_CMD_TEXT\")"
+            )
             PsiDocumentManager.getInstance(project).commitDocument(doc)
         }
     }
@@ -207,7 +231,7 @@ class SmkUnusedLogFileInspection : SnakemakeInspection() {
             val indent = lastArg?.getPreviousOffset() ?: logIndent
             doc.insertString(
                 (lastArg ?: ruleLike).endOffset,
-                "${indent}shell: \"echo TODO$REDIRECT_STDERR_STDOUT_TO_LOG_CMD_TEXT\""
+                "${indent}shell: \"(echo TODO) $REDIRECT_STDERR_STDOUT_TO_LOG_CMD_TEXT\""
             )
             PsiDocumentManager.getInstance(project).commitDocument(doc)
         }